Clear transaction state on tab changes
